Executable version 1
Fixed all of the bugs, including the one where the snakes could sometimes pass through walls.
Snakes are now told where the fruit are located and are capable of pathing to them.
There are still no iterations, that will be the next task.
Press any button to reset the game.